+++ src/service-worker.js
@@ -112,8 +112,8 @@ export default class ServiceWorker {
           optional: cache('optional'),
         },
         strategy: plugin.strategy,
-        version: plugin.strategy === 'all' ? plugin.version : void 0,
-        hash: plugin.strategy !== 'all' ? plugin.hash : void 0,
+        version: plugin.strategy !== 'hash' ? plugin.version : void 0,
+        hash: plugin.strategy === 'hash' ? plugin.hash : void 0,
         name: this.CACHE_NAME,
         relativePaths: plugin.relativePaths
       }, null, minify ? void 0 : '  ') };