+++ CHANGELOG.md
@@ -14,6 +14,12 @@
 * Cache sections `'additional'` and `'optional'` are now allowed only when `updateStrategy`option is set to `'changed'`
 * `updateStrategy: `changed`` now uses `version` option instead of compilation hash.
 * Version now is not set by default and returns compilation hash for `updateStrategy: 'changed`` and `version` for `updateStrategy: 'all'` (when not set)
+* Added `alwaysRevalidate` option
+* Improve `updateStrategy` option implementation. `changed` mode might be still unstable--**do not use it yet**
+* `install()` method signature now is `install(options)` (callbacks are removed)
+* Runtime events are not implemented for ServiceWorker: `onUpdating`, `onUpdateReady`, `onUpdated`, `onError`.  
+  Example: `runtime.install({ onError: () => ... })`
+* Added `applyUpdate()` method of runtime
 
 ### 2.0.0